PartiallySeparable.jl Documentation

test modif

PartiallySeparableStructure.deduct_partially_separable_structure — Method

deduct_partially_separable_structure(expr_tree, n)

Find the partially separable structure of a function f stored as an expression tree expr_tree. To define properly the size of sparse matrix we need the size of the problem : n. At the end, we get the partially separable structure of f, f(x) = ∑fᵢ(xᵢ)

PartiallySeparableStructure.build_gradient — Method

build_gradient(sps, g)

Constructs a vector of size n from the list of element gradient of the sps structure which has numerous element gradient of size nᵢ. The purpose of the function is to gather these element gradient into a real gradient of size n. The function gradnito_n! will add element gradient of size nᵢ at the right inside the gradient of size n.

PartiallySeparableStructure.compiled_grad_of_elmt_fun — Method

compiled_grad_of_elmt_fun(elmt_fun)

Return the GradientTape compiled to speed up the ReverseDiff computation of the elmt_fun gradient in the future

PartiallySeparableStructure.construct_Sparse_Hessian — Method

constructSparseHessian(sps, B) Build from the Partially separable Structure sps and the Hessian approximation B a SpaseArray which represent B in other form.

PartiallySeparableStructure.element_gradient! — Method

elementgradient!(compiltape, x, g) Compute the element grandient from the compil_tape compiled before according to the vector x, and store the result in the vector g Use of ReverseDiff

PartiallySeparableStructure.element_gradient2! — Method

element_gradient2!(expr_tree, x, g)

Compute the element grandient of the function represents by expr_tree according to the vector x, and store the result in the vector g. This was the previous version using ForwardDiff. The actual version using ReverseDiff is more efficient.

PartiallySeparableStructure.evaluate_SPS — Method

evaluate_SPS(sps,x)

evalutate the partially separable function f = ∑fᵢ, stored in the sps structure at the point x. f(x) = ∑fᵢ(xᵢ), so we compute independently each fᵢ(xᵢ) and we return the sum.

PartiallySeparableStructure.evaluate_SPS_gradient! — Method

evaluate_SPS_gradient!(sps,x,g)

Compute the gradient of the partially separable structure sps, and store the result in the gradvector structure g. Using ReversDiff package. Not obvious good behaviour with Threads.@threads, sometime yes sometime no. Noted that we use the previously compiled GradientTape in elementgradient! that use ReverseDiff.

PartiallySeparableStructure.evaluate_SPS_gradient2! — Method

evaluate_SPS_gradient2!(sps,x,g)

Compute the gradient of the partially separable structure sps, and store the result in the grad_vector structure g. Using ForwardDiff package. Bad behaviour with Threads.@threads. This was the previous version using ForwardDiff. The actual version using ReverseDiff is more efficient.

PartiallySeparableStructure.evaluate_element_hessian — Method

evaluate_element_hessian(fᵢ,xᵢ)

Compute the Hessian of the elemental function fᵢ : Gᵢ a n × n matrix. So xᵢ a vector of size nᵢ. The result of the function is the triplet of the sparse matrix Gᵢ.

PartiallySeparableStructure.evaluate_gradient — Method

evaluate_gradient(sps,x)

evalutate the gradient of the partially separable function f = ∑ fι, stored in the sps structure at the point x, return a vector of size n (the number of variable) which is the gradient. Première version de la fonction inutile car inefficace.

PartiallySeparableStructure.evaluate_hessian — Method

evaluate_hessian(sps,x)

evalutate the hessian of the partially separable function f = ∑ fᵢ, stored in the sps structure at the point x. Return the sparse matrix of the hessian of size n × n.

PartiallySeparableStructure.id_hessian! — Method

id_hessian!(sps, B)

Construct a kinf of Id Hessian, it will initialize each element Hessian Bᵢ with an Id matrix, B = ∑ᵢᵐ Uᵢᵀ Bᵢ Uᵢ

PartiallySeparableStructure.minus_grad_vec! — Method

minus_grad_vec!(g1,g2,res)

Store in res: g1 minus g2, but g1 and g2 have a particular structure which is grad_vector{T}. We need this operation to have the difference for each element gradient for TR method. g1 = gₖ and g2 = gₖ₋₁.

PartiallySeparableStructure.product_matrix_sps! — Method

product_matrix_sps!(sps,B,x,Bx)

This function make the product of the structure B which represents a symetric matrix and the vector x, the result is stored in Bx. We need the structure sps for the variable used in each B[i], to replace B[i]*x[i] in the result vector by using f_inter!.

PartiallySeparableStructure.product_matrix_sps — Method

product_matrix_sps(sps,B,x)

This function make the product of the structure B which represents a symetric matrix and the vector x. We need the structure sps for the variable used in each B[i], to replace B[i]*x[i] in the result vector.

PartiallySeparableStructure.product_vector_sps — Method

product_vector_sps(sps, g, x)

compute the product g⊤ x = ∑ Uᵢ⊤ gᵢ⊤ xᵢ. So we need the sps structure to get the Uᵢ. On ne s'en sert pas en pratique mais peut-être pratique pour faire des vérifications

PartiallySeparableStructure.struct_hessian! — Method

struct_hessian!(sps,x,H)

Evalutate the hessian of the partially separable function, stored in the sps structure at the point x. Store the Hessian in a particular structure H :: Hess_matrix.

PartiallySeparableStructure.struct_hessian — Method

struct_hessian(sps,x)

evalutate the hessian of the partially separable function, stored in the sps structure at the point x. Return the Hessian in a particular structure : Hess_matrix.

PartiallySeparableStructure.update_SPS_BFGS! — Method

update_SPS_BFGS(sps, Bₖ, Bₖ₊₁, yₖ, sₖ)

update the Hessian approximation Bₖ using the SR1 method, according to the sps partially separable structre. To make the update, we need the gradvector y and the vector s. B, B1 and y use structure linked with the partially separable structure stored in sps.

PartiallySeparableStructure.update_SPS_SR1! — Method

update_SPS_SR1(sps, Bₖ, Bₖ₊₁, yₖ, sₖ)

update the Hessian approximation Bₖ using the SR1 method, according to the sps partially separable structre. To make the update, we need the vector y and s.

PartiallySeparableStructure.update_SPS_SR1! — Method

update_SPS_SR1(sps, Bₖ, Bₖ₊₁, yₖ, sₖ)

update the Hessian approximation Bₖ using the SR1 method, according to the sps partially separable structre. To make the update, we need the gradvector y and the vector s. B, B1 and y use structure linked with the partially separable structure stored in sps.